feat: store assertion message and stack trace separately
The CTRF Test objects has a message and a trace attribute. According to
the documentation, the message should contain a description of the test
result and the trace should contain the stack trace of a failed test.

The reporter receives the assertion description and the stack trace
intermingled and stores them both, redundantly, in the message and trace
attributes.

This change splits the assert description from the stack trace as shown
stores them separately. Everything above the dashed line is stored in
message while lines below are stored in trace.

```
Error: expect(received).toStrictEqual(expected) // deep equality

- Expected  - 1
+ Received  + 1

@@ -23,11 +23,11 @@
          "isSelected": true,
        },
      ],
      "damper_setup": "damper_1,damper_2",
      "driver": "driver1",
-     "id": 0,
+     "id": 1,
      "isCurrentlyUsedInProfile": true,
      "isIntermediateState": true,
      "isSelectedForAll": false,
      "lap": "LapName1",
      "lap_id": 1,
-------------------------
    at lap-channel-type-table-data-creation.service.spec.ts:185:26
    at zone-testing-bundle.umd.js:409:30)
    at zone-testing-bundle.umd.js:3830:43)
    at zone.js/bundles/zone-testing-bundle.umd.js:408:56)
```

This simplifies processing of both values for later tools when eval-
uating the test results.